Create ASCII_SYMBOL unordered map
We are now using ordinal numbers for the keys because the app is no longer using the char() function to generate an ASCII character.

Also, I did the initial performance test for this version. The result is not bad with an average performance improvement of 10.97% generating a 100,000-length string. I'm now impressed and a newly minted fan of unordered_map 👌.